Refactoring: component separation
