+++ src/Counter.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-function Counter({ count, onClick }) {
+export default function Counter({ count, onClick }) {
   return (
     <>
       <button type="button" onClick={() => onClick(1)}>
@@ -12,5 +12,3 @@ function Counter({ count, onClick }) {
     </>
   );
 }
-
-export default Counter;
