+++ src/Page.jsx
@@ -1,17 +1,14 @@
 import React from 'react';
+
 import Counter from './Counter';
 import Buttons from './Buttons';
 
-function Page({ count, onClick }) {
+export default function Page({ count, onClick }) {
   return (
     <div>
       <p>Counter</p>
       <Counter count={count} onClick={onClick} />
-      <p>
-        <Buttons count={count} onClick={onClick} />
-      </p>
+      <Buttons count={count} onClick={onClick} />
     </div>
   );
 }
-
-export default Page;
