+++ src/index.jsx
@@ -1,23 +1,7 @@
 /* @jsx React.createElement */
 
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import Page from './Page';
-
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-
-  const { count } = state;
-
-  function handleClick(number) {
-    setState({
-      count: count + number,
-    });
-  }
-
-  return <Page count={count} onClick={handleClick} />;
-}
+import App from './App';
 
 ReactDOM.render(<App />, document.getElementById('app'));
